moved `PCStateDiagnostics` to `diagnostics` package
